[hash] Move sha256 module out of compress.rs
Extract the inline pub mod sha256 { ... } block from compress.rs into its
own top-level crates/hash/src/sha256.rs, exposed at binius_hash::sha256.
The StdCompression type alias is repointed at sha256::Sha256Compression.

This makes the module structure consistent with the upcoming HashSuite
trait, which will define a Sha256HashSuite (and a parallel VisionHashSuite)
each living next to its companion compression function.
