+++ crates/hash/src/lib.rs
@@ -13,6 +13,7 @@ pub mod compress;
 pub mod parallel_compression;
 pub mod parallel_digest;
 mod serialization;
+pub mod sha256;
 pub mod vision;
 pub mod vision_4;
 pub mod vision_6;
@@ -24,4 +25,4 @@ pub use serialization::*;
 
 /// The standard digest is SHA-256.
 pub type StdDigest = sha2::Sha256;
-pub type StdCompression = compress::sha256::Sha256Compression;
+pub type StdCompression = sha256::Sha256Compression;
